added Prevayler support for Account

+++ db-bench/src/main/java/com/eris4/benchdb/database/pico4/account/AccountPico4Driver.java
@@ -41,7 +41,7 @@ public class AccountPico4Driver implements AccountDriver {
 	}
 
 	@Override
-	public Account read(long accountId) throws TestDriverException {
+	public Account read(int accountId) throws TestDriverException {
 		return (Account) map.get(String.valueOf(accountId));
 	}
